refactor(state): split legacy migration helpers
