+++ extensions/memory-core/doctor-contract-api.ts
@@ -4,7 +4,6 @@ import fsSync from "node:fs";
 import fs from "node:fs/promises";
 import path from "node:path";
 import type { DatabaseSync } from "node:sqlite";
-import { isDeepStrictEqual } from "node:util";
 import { resolveUserPath } from "openclaw/plugin-sdk/memory-core-host-engine-foundation";
 import {
   ensureMemoryIndexSchema,
@@ -47,6 +46,7 @@ import {
   writeMemoryCoreWorkspaceEntries,
   writeMemoryCoreWorkspaceEntry,
 } from "./src/dreaming-state.js";
+import { dreamingStateComparison } from "./src/migration/dreaming-state-comparison.js";
 import {
   SHORT_TERM_PHASE_SIGNAL_RELATIVE_PATH,
   SHORT_TERM_STORE_RELATIVE_PATH,
@@ -1114,111 +1114,6 @@ async function migratePhaseSignals(source: LegacySource): Promise<number> {
   return Object.keys(state.entries).length;
 }
 
-function targetNamespacesForSource(label: string): string[] {
-  if (label === "daily ingestion") {
-    return [DREAMING_DAILY_INGESTION_NAMESPACE];
-  }
-  if (label === "session ingestion") {
-    return [DREAMING_SESSION_INGESTION_FILES_NAMESPACE, DREAMING_SESSION_INGESTION_SEEN_NAMESPACE];
-  }
-  if (label === "short-term recall") {
-    return [SHORT_TERM_RECALL_NAMESPACE];
-  }
-  return [SHORT_TERM_PHASE_SIGNAL_NAMESPACE];
-}
-
-async function legacySourceMatchesCanonical(source: LegacySource): Promise<boolean> {
-  const raw = await readJsonFile(source.filePath);
-  if (source.label === "daily ingestion") {
-    const rows = await readMemoryCoreWorkspaceEntries({
-      namespace: DREAMING_DAILY_INGESTION_NAMESPACE,
-      workspaceDir: source.workspaceDir,
-    });
-    return isDeepStrictEqual(
-      normalizeDailyIngestionState(raw),
-      normalizeDailyIngestionState({
-        version: 1,
-        files: Object.fromEntries(rows.map((row) => [row.key, row.value])),
-      }),
-    );
-  }
-  if (source.label === "session ingestion") {
-    const [fileRows, seenRows] = await Promise.all([
-      readMemoryCoreWorkspaceEntries({
-        namespace: DREAMING_SESSION_INGESTION_FILES_NAMESPACE,
-        workspaceDir: source.workspaceDir,
-      }),
-      readMemoryCoreWorkspaceEntries<{ scope: string; index: number; hashes: string[] }>({
-        namespace: DREAMING_SESSION_INGESTION_SEEN_NAMESPACE,
-        workspaceDir: source.workspaceDir,
-      }),
-    ]);
-    const chunksByScope = new Map<string, Array<{ index: number; hashes: string[] }>>();
-    for (const row of seenRows) {
-      const chunks = chunksByScope.get(row.value.scope) ?? [];
-      chunks.push({ index: row.value.index, hashes: row.value.hashes });
-      chunksByScope.set(row.value.scope, chunks);
-    }
-    return isDeepStrictEqual(
-      normalizeSessionIngestionState(raw),
-      normalizeSessionIngestionState({
-        version: 3,
-        files: Object.fromEntries(fileRows.map((row) => [row.key, row.value])),
-        seenMessages: Object.fromEntries(
-          [...chunksByScope].map(([scope, chunks]) => [
-            scope,
-            chunks.toSorted((left, right) => left.index - right.index).flatMap((row) => row.hashes),
-          ]),
-        ),
-      }),
-    );
-  }
-  const [entryRows, metaRows] = await Promise.all([
-    readMemoryCoreWorkspaceEntries({
-      namespace:
-        source.label === "short-term recall"
-          ? SHORT_TERM_RECALL_NAMESPACE
-          : SHORT_TERM_PHASE_SIGNAL_NAMESPACE,
-      workspaceDir: source.workspaceDir,
-    }),
-    readMemoryCoreWorkspaceEntries<{ updatedAt?: unknown }>({
-      namespace: SHORT_TERM_META_NAMESPACE,
-      workspaceDir: source.workspaceDir,
-    }),
-  ]);
-  const metaKey = source.label === "short-term recall" ? "recall" : "phase";
-  const updatedAt = metaRows.find((row) => row.key === metaKey)?.value.updatedAt;
-  if (typeof updatedAt !== "string" || updatedAt.length === 0) {
-    return false;
-  }
-  const canonicalRaw = {
-    version: 1,
-    updatedAt,
-    entries: Object.fromEntries(entryRows.map((row) => [row.key, row.value])),
-  };
-  if (source.label === "short-term recall") {
-    const canonical = normalizeShortTermRecallStore(canonicalRaw, updatedAt);
-    const fallbackCandidates = new Set([updatedAt]);
-    for (const row of entryRows) {
-      const value = asRecord(row.value);
-      for (const key of ["firstRecalledAt", "lastRecalledAt"] as const) {
-        if (typeof value?.[key] === "string") {
-          fallbackCandidates.add(value[key]);
-        }
-      }
-    }
-    // Legacy entries may omit recall timestamps. Migration filled every
-    // omission with one migration-time value, which survives only in rows.
-    return [...fallbackCandidates].some((fallback) =>
-      isDeepStrictEqual(normalizeShortTermRecallStore(raw, fallback), canonical),
-    );
-  }
-  return isDeepStrictEqual(
-    normalizeShortTermPhaseSignalStore(raw, updatedAt),
-    normalizeShortTermPhaseSignalStore(canonicalRaw, updatedAt),
-  );
-}
-
 async function migrateSource(source: LegacySource): Promise<number> {
   if (source.label === "daily ingestion") {
     return await migrateDailyIngestion(source);
@@ -1254,23 +1149,11 @@ export const stateMigrations: PluginDoctorStateMigration[] = [
       const warnings: string[] = [];
       const notices: string[] = [];
       for (const source of await collectLegacySources(params.config, params.env)) {
-        const targetHasRows = (
-          await Promise.all(
-            targetNamespacesForSource(source.label).map(
-              async (namespace) =>
-                (
-                  await readMemoryCoreWorkspaceEntries({
-                    namespace,
-                    workspaceDir: source.workspaceDir,
-                  })
-                ).length > 0,
-            ),
-          )
-        ).some(Boolean);
+        const targetHasRows = await dreamingStateComparison.targetHasRows(source);
         if (targetHasRows) {
           let matchesCanonical = false;
           try {
-            matchesCanonical = await legacySourceMatchesCanonical(source);
+            matchesCanonical = await dreamingStateComparison.sourceMatchesCanonical(source);
           } catch (err) {
             warnings.push(
               `Skipped Memory Core ${source.label} import for ${source.workspaceDir} because the legacy source could not be compared: ${String(err)}`,
